Refactor TransferClient to use pydantic types
