+++ src/pypaystack2/paystack.py
@@ -25,6 +25,7 @@ from pypaystack2.sub_clients.async_clients.transactions import AsyncTransactionC
 from pypaystack2.sub_clients.async_clients.transfer_recipients import (
     AsyncTransferRecipientClient,
 )
+from pypaystack2.sub_clients.async_clients.transfers import AsyncTransferClient
 from pypaystack2.sub_clients.sync_clients.apple_pay import (
     ApplePayClient,
 )
@@ -79,7 +80,9 @@ from pypaystack2.sub_clients.sync_clients.transactions import (
 from pypaystack2.sub_clients.sync_clients.transfer_recipients import (
     TransferRecipientClient,
 )
-from pypaystack2.sub_clients.transfers import AsyncTransferClient, TransferClient
+from pypaystack2.sub_clients.sync_clients.transfers import (
+    TransferClient,
+)
 from pypaystack2.sub_clients.transfers_control import (
     AsyncTransferControlClient,
     TransferControlClient,
