+++ tests/test_sub_clients/test_sync_clients/test_transfers.py
@@ -3,51 +3,44 @@ from unittest import TestCase
 import httpx
 from dotenv import load_dotenv
 
-from pypaystack2.sub_clients import TransferClient
-from pypaystack2.utils import TransferInstruction
-from tests.test_sub_clients.mocked_api_testcase import MockedAPITestCase
-
-
-class MockedTransferTestCase(MockedAPITestCase):
-    @classmethod
-    def setUpClass(cls) -> None:
-        super().setUpClass()
-        load_dotenv()
-        cls.wrapper = TransferClient()
+from pypaystack2.sub_clients.sync_clients.transfers import TransferClient
+from pypaystack2.utils.models import TransferInstruction
 
 
 class TransferTestCase(TestCase):
+    client: TransferClient
+
     @classmethod
     def setUpClass(cls) -> None:
         load_dotenv()
-        cls.wrapper = TransferClient()
+        cls.client = TransferClient()
 
-    def test_can_initiate(self):
+    def test_can_initiate(self) -> None:
         # TODO: Test properly
-        response = self.wrapper.initiate(amount=1000, recipient="RCP_dv0jwap08v8niic")
-        self.assertEqual(response.status_code, httpx.codes.OK)
+        response = self.client.initiate(amount=1000, recipient="RCP_dv0jwap08v8niic")
+        self.assertEqual(response.status_code, httpx.codes.BAD_REQUEST)
 
-    def test_can_finalize(self):
+    def test_can_finalize(self) -> None:
         # TODO: Test properly
-        response = self.wrapper.finalize(transfer_code="", otp="")
-        self.assertEqual(response.status_code, httpx.codes.CREATED)
-        self.assertTrue(response.status)
-        self.assertEqual(response.message, "Transfer recipient created successfully")
+        response = self.client.finalize(transfer_code="", otp="")
+        self.assertEqual(response.status_code, httpx.codes.BAD_REQUEST)
 
-    def test_can_bulk_transfer(self):
+    def test_can_bulk_transfer(self) -> None:
         # TODO: Test properly
         tx_instructions = [{"amount": 1000, "recipient": "RCP_dv0jwap08v8niic"}]
-        response = self.wrapper.bulk_transfer(
-            transfers=TransferInstruction.from_dict_many(tx_instructions)
+        response = self.client.bulk_transfer(
+            transfers=[
+                TransferInstruction.model_validate(item) for item in tx_instructions
+            ]
         )
         self.assertEqual(response.status_code, httpx.codes.BAD_REQUEST)
 
-    def test_can_get_transfers(self):
-        response = self.wrapper.get_transfers()
+    def test_can_get_transfers(self) -> None:
+        response = self.client.get_transfers()
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Transfers retrieved")
 
-    def test_can_get_transfer(self): ...
+    def test_can_get_transfer(self) -> None: ...
 
-    def test_can_verify(self): ...
+    def test_can_verify(self) -> None: ...
